pyproject.toml
[project]
name = "simpletemplate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "sqlmodel>=0.0.24",
    "python-dotenv>=1.0.0",
    "alembic>=1.13.1",
    "pytest>=7.4.0",
    "httpx>=0.25.0",  # 用于测试
]

[project.optional-dependencies]
dev = [
    "black>=23.7.0",
    "isort>=5.12.0",
    "mypy>=1.5.1",
    "pytest-cov>=4.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
